refactor: move operations with files to separate module
